Split out GCP client
